experimental changes

+++ quill/include/quill/Config.h
@@ -180,6 +180,6 @@ struct Config
    *
    * @note This capacity automatically doubles when the unbounded queue is full
    */
-  size_t default_queue_capacity{131'072};
+  int32_t default_queue_capacity{131'072};
 };
 } // namespace quill
