+++ quill/include/quill/Logger.h
@@ -142,8 +142,9 @@ public:
     size_t c_string_sizes[(std::max)(c_string_count, static_cast<size_t>(1))];
 
     // Need to reserve additional space as we will be aligning the pointer
-    size_t const total_size = sizeof(detail::Header) + alignof(detail::Header) +
-      detail::get_args_sizes<0>(c_string_sizes, fmt_args...);
+    int32_t const total_size = static_cast<int32_t>(sizeof(detail::Header)) +
+      static_cast<int32_t>(alignof(detail::Header)) +
+      static_cast<int32_t>(detail::get_args_sizes<0>(c_string_sizes, fmt_args...));
 
     // request this size from the queue
     std::byte* write_buffer = thread_context->spsc_queue().prepare_write(total_size);
@@ -178,11 +179,11 @@ public:
 
     // encode remaining arguments
     write_buffer = detail::encode_args<0>(c_string_sizes, write_buffer, std::forward<FmtArgs>(fmt_args)...);
-    assert(total_size >= (static_cast<size_t>(write_buffer - write_begin)) &&
+    assert(total_size >= (static_cast<int32_t>(write_buffer - write_begin)) &&
            "The committed write bytes can not be greater than the requested bytes");
     assert((write_buffer >= write_begin) &&
            "write_buffer should be greater or equal to write_begin");
-    thread_context->spsc_queue().commit_write(static_cast<size_t>(write_buffer - write_begin));
+    thread_context->spsc_queue().commit_write(static_cast<int32_t>(write_buffer - write_begin));
   }
 
   /**
