+++ quill/include/quill/detail/ThreadContext.h
@@ -36,7 +36,7 @@ public:
   /**
    * Constructor
    */
-  explicit ThreadContext(size_t default_queue_capacity) : _spsc_queue(default_queue_capacity) {}
+  explicit ThreadContext(int32_t default_queue_capacity) : _spsc_queue(default_queue_capacity) {}
 
   /**
    * Deleted
